# 07_API_CONTRACT.yaml
openapi: 3.1.0
info:
  title: Workshops Platform API
  version: 1.0.0
  description: >-
    العقد المرجعي لمنصة ورش. المصادقة نفسها تحت /api/auth/* تديرها Better Auth؛
    بقية المسارات تحت /api/v1 وتستخدم جلسة Cookie آمنة وCSRF عند تغيير الحالة.
servers:
  - url: https://workshops.gumaan.app/api/v1
    description: Production
  - url: http://localhost:5173/api/v1
    description: Local
...
security:
  - sessionCookie: []
paths:
...
  /workspaces/{workspaceId}/organizations:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Organizations]
      operationId: listOrganizations
      x-permission: organization.read
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: primaryField
          in: query
          schema: {type: string}
        - name: region
          in: query
          schema: {type: string}
      responses:
        "200": {$ref: "#/components/responses/OrganizationList"}
    post:
      tags: [Organizations]
      operationId: createOrganization
      x-permission: organization.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/OrganizationInput"}
      responses:
        "201": {$ref: "#/components/responses/OrganizationItem"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
  /workspaces/{workspaceId}/organizations/{organizationId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Organizations]
      operationId: getOrganization
      x-permission: organization.read
      responses:
        "200": {$ref: "#/components/responses/OrganizationItem"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      x-permission: organization.manage
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/OrganizationPatch"}
      responses:
        "200": {$ref: "#/components/responses/OrganizationItem"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /workspaces/{workspaceId}/workshops:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Workshops]
      operationId: listWorkshops
      x-permission: workshop.read
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema: {$ref: "#/components/schemas/WorkshopStatus"}
        - name: programId
          in: query
          schema: {$ref: "#/components/schemas/Id"}
      responses:
        "200": {$ref: "#/components/responses/WorkshopList"}
    post:
      tags: [Workshops]
      operationId: createWorkshop
      x-permission: workshop.create
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/WorkshopInput"
                - type: object
                  properties:
                    source:
                      type: object
                      properties:
                        type: {type: string, enum: [blank, template, workshop_copy]}
                        id: {$ref: "#/components/schemas/Id"}
      responses:
        "201": {$ref: "#/components/responses/WorkshopItem"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /workspaces/{workspaceId}/workshops/{workshopId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/WorkshopId"
    get:
      tags: [Workshops]
      operationId: getWorkshop
      x-permission: workshop.read
      responses:
        "200": {$ref: "#/components/responses/WorkshopItem"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [Workshops]
      operationId: updateWorkshop
      x-permission: workshop.update
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/WorkshopPatch"}
      responses:
        "200": {$ref: "#/components/responses/WorkshopItem"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /workspaces/{workspaceId}/workshops/{workshopId}/attendance:
    get:
      tags: [Registration]
      operationId: listAttendance
      x-permission: attendance.manage or report.read
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/WorkshopId"
        - name: dayId
          in: query
          schema: {$ref: "#/components/schemas/Id"}
      responses:
        "200":
          description: Attendance matrix.
    post:
      tags: [Registration]
      operationId: recordAttendance
      x-permission: attendance.manage
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/WorkshopId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AttendanceInput"}
      responses:
        "200":
          description: Existing or newly recorded attendance.
        "422": {$ref: "#/components/responses/ValidationError"}
...
  /workspaces/{workspaceId}/workshops/{workshopId}/activities:
    get:
      tags: [Activities]
      operationId: listActivities
      x-permission: workshop.read
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/WorkshopId"
        - name: status
          in: query
          schema: {$ref: "#/components/schemas/ActivityStatus"}
      responses:
        "200": {$ref: "#/components/responses/ActivityList"}
    post:
      tags: [Activities]
      operationId: createActivity
      x-permission: activity.design
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/WorkshopId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ActivityInput"}
      responses:
        "201": {$ref: "#/components/responses/ActivityItem"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /workspaces/{workspaceId}/workshops/{workshopId}/activities/{activityId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/WorkshopId"
      - $ref: "#/components/parameters/ActivityId"
    get:
      tags: [Activities]
      operationId: getActivity
      x-permission: workshop.read or assigned participant
      responses:
        "200": {$ref: "#/components/responses/ActivityItem"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [Activities]
      operationId: updateActivity
      x-permission: activity.design
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ActivityPatch"}
      responses:
        "200": {$ref: "#/components/responses/ActivityItem"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-workshops_session
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Id"}
    WorkshopId:
      name: workshopId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Id"}
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Id"}
    ActivityId:
      name: activityId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Id"}
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 25}
    Search:
      name: q
      in: query
      schema: {type: string, maxLength: 200}
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string}
      description: Entity version or ETag used for optimistic concurrency.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
  schemas:
    Id:
      type: string
      minLength: 8
      maxLength: 64
      pattern: "^[A-Za-z0-9_-]+$"
    Timestamp:
      type: string
      format: date-time
    Error:
      type: object
      required: [code, message, requestId]
      additionalProperties: false
      properties:
        code: {type: string}
        message: {type: string}
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items: {type: string}
        requestId: {type: string}
        details: {type: object, additionalProperties: true}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: {$ref: "#/components/schemas/Error"}
    PageMeta:
      type: object
      required: [hasMore]
      properties:
        nextCursor: {type: [string, "null"]}
        hasMore: {type: boolean}
        total: {type: integer, minimum: 0}
...
    Organization:
      type: object
      required: [id, name, status, version]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        name: {type: string}
        shortName: {type: [string, "null"]}
        region: {type: [string, "null"]}
        city: {type: [string, "null"]}
        primaryField: {type: [string, "null"]}
        secondaryFields: {type: array, items: {type: string}}
        description: {type: [string, "null"]}
        website: {type: [string, "null"], format: uri}
        status: {type: string, enum: [active, archived, merged]}
        version: {type: integer, minimum: 1}
    OrganizationInput:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: {type: string, minLength: 2, maxLength: 200}
        shortName: {type: string, maxLength: 80}
        region: {type: string, maxLength: 120}
        city: {type: string, maxLength: 120}
        primaryField: {type: string, maxLength: 160}
        secondaryFields: {type: array, maxItems: 30, items: {type: string, maxLength: 120}}
        description: {type: string, maxLength: 4000}
        website: {type: string, format: uri}
        contactName: {type: string, maxLength: 160}
        contactEmail: {type: string, format: email}
        contactPhone: {type: string, maxLength: 40}
    OrganizationPatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name: {type: string, minLength: 2, maxLength: 200}
        shortName: {type: string, maxLength: 80}
        region: {type: string, maxLength: 120}
        city: {type: string, maxLength: 120}
        primaryField: {type: string, maxLength: 160}
        secondaryFields: {type: array, maxItems: 30, items: {type: string, maxLength: 120}}
        description: {type: string, maxLength: 4000}
        website: {type: string, format: uri}
        contactName: {type: string, maxLength: 160}
        contactEmail: {type: string, format: email}
        contactPhone: {type: string, maxLength: 40}
        status: {type: string, enum: [active, archived]}
...
    WorkshopStatus:
      type: string
      enum: [draft, registration_open, scheduled, live, paused, completed, archived]
    Workshop:
      type: object
      required: [id, slug, title, deliveryMode, timezone, status, visibility, version]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        programId: {anyOf: [{$ref: "#/components/schemas/Id"}, {type: "null"}]}
        slug: {type: string}
        title: {type: string}
        subtitle: {type: [string, "null"]}
        description: {type: [string, "null"]}
        deliveryMode: {type: string, enum: [in_person, online, hybrid]}
        venueName: {type: [string, "null"]}
        meetingUrl: {type: [string, "null"], format: uri}
        timezone: {type: string}
        startAt: {anyOf: [{$ref: "#/components/schemas/Timestamp"}, {type: "null"}]}
        endAt: {anyOf: [{$ref: "#/components/schemas/Timestamp"}, {type: "null"}]}
        capacity: {type: [integer, "null"], minimum: 1}
        registrationMode: {type: string, enum: [closed, invite_only, public]}
        registrationApproval: {type: string, enum: [automatic, manual]}
        status: {$ref: "#/components/schemas/WorkshopStatus"}
        visibility: {type: string, enum: [private, unlisted, public]}
        version: {type: integer, minimum: 1}
    WorkshopInput:
      type: object
      required: [slug, title, deliveryMode, timezone]
      additionalProperties: false
      properties:
        programId: {$ref: "#/components/schemas/Id"}
        slug: {type: string, pattern: "^[a-z0-9-]{3,80}$"}
        title: {type: string, minLength: 2, maxLength: 240}
        subtitle: {type: string, maxLength: 300}
        description: {type: string, maxLength: 10000}
        deliveryMode: {type: string, enum: [in_person, online, hybrid]}
        venueName: {type: string, maxLength: 240}
        venueAddress: {type: string, maxLength: 1000}
        meetingUrl: {type: string, format: uri}
        timezone: {type: string}
        startAt: {$ref: "#/components/schemas/Timestamp"}
        endAt: {$ref: "#/components/schemas/Timestamp"}
        capacity: {type: integer, minimum: 1}
        registrationMode: {type: string, enum: [closed, invite_only, public], default: invite_only}
        registrationApproval: {type: string, enum: [automatic, manual], default: manual}
        visibility: {type: string, enum: [private, unlisted, public], default: private}
    WorkshopPatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        programId: {$ref: "#/components/schemas/Id"}
        slug: {type: string, pattern: "^[a-z0-9-]{3,80}$"}
        title: {type: string, minLength: 2, maxLength: 240}
        subtitle: {type: string, maxLength: 300}
        description: {type: string, maxLength: 10000}
        deliveryMode: {type: string, enum: [in_person, online, hybrid]}
        venueName: {type: string, maxLength: 240}
        venueAddress: {type: string, maxLength: 1000}
        meetingUrl: {type: string, format: uri}
        timezone: {type: string}
        startAt: {$ref: "#/components/schemas/Timestamp"}
        endAt: {$ref: "#/components/schemas/Timestamp"}
        capacity: {type: integer, minimum: 1}
        registrationMode: {type: string, enum: [closed, invite_only, public]}
        registrationApproval: {type: string, enum: [automatic, manual]}
        visibility: {type: string, enum: [private, unlisted, public]}
...
    AttendanceInput:
      type: object
      required: [workshopDayId, workshopMembershipId, status, method]
      properties:
        workshopDayId: {$ref: "#/components/schemas/Id"}
        agendaItemId: {$ref: "#/components/schemas/Id"}
        workshopMembershipId: {$ref: "#/components/schemas/Id"}
        status: {type: string, enum: [present, late, excused, absent]}
        method: {type: string, enum: [manual, qr, code, import]}
        checkInCode: {type: string}
        correctionReason: {type: string, maxLength: 1000}
...
    ActivityStatus:
      type: string
      enum: [draft, published, open, paused, closed, archived]
    ActivityFieldInput:
      type: object
      required: [key, type, label]
      additionalProperties: false
      properties:
        key: {type: string, pattern: "^[a-z][a-z0-9_]{1,63}$"}
        type: {type: string, enum: [short_text, long_text, number, scale, single_choice, multi_choice, boolean, ranking, matrix, date, time, url, file]}
        label: {type: string, minLength: 1, maxLength: 500}
        description: {type: string, maxLength: 2000}
        placeholder: {type: string, maxLength: 500}
        required: {type: boolean, default: false}
        sensitive: {type: boolean, default: false}
        comparisonKey: {type: string, maxLength: 100}
        validation: {type: object, additionalProperties: true}
        visibilityRule: {type: object, additionalProperties: true}
        scoring: {type: object, additionalProperties: true}
        options:
          type: array
          items:
            type: object
            required: [key, label]
            properties:
              key: {type: string}
              label: {type: string}
              value: true
    ActivityInput:
      type: object
      required: [type, title, ownershipMode, identityMode, fields]
      additionalProperties: false
      properties:
        type: {type: string, enum: [survey, quiz, poll, individual_exercise, group_exercise, self_assessment, facilitator_assessment, commitment, form]}
        title: {type: string, minLength: 1, maxLength: 240}
        purpose: {type: string, maxLength: 2000}
        instructions: {type: string, maxLength: 10000}
        expectedOutput: {type: string, maxLength: 4000}
        ownershipMode: {type: string, enum: [individual, organization, group, facilitator]}
        identityMode: {type: string, enum: [named, facilitator_hidden, anonymous]}
        durationMinutes: {type: integer, minimum: 1, maximum: 1440}
        attemptPolicy: {type: string, enum: [single, multiple, editable_until_close]}
        showResultsPolicy: {type: string, enum: [never, after_submit, after_close, facilitator_controlled]}
        comparisonPhase: {type: string, enum: [pre, post]}
        fields: {type: array, minItems: 1, items: {$ref: "#/components/schemas/ActivityFieldInput"}}
        targets:
          type: array
          minItems: 1
          items:
            type: object
            required: [type]
            properties:
              type: {type: string, enum: [all_participants, organization, group, user]}
              id: {$ref: "#/components/schemas/Id"}
...
    ActivityPatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        title: {type: string, minLength: 1, maxLength: 240}
        purpose: {type: string, maxLength: 2000}
        instructions: {type: string, maxLength: 10000}
        expectedOutput: {type: string, maxLength: 4000}
        durationMinutes: {type: integer, minimum: 1, maximum: 1440}
        fields: {type: array, minItems: 1, items: {$ref: "#/components/schemas/ActivityFieldInput"}}
...
    OrganizationEnvelope:
      type: object
      properties:
        data: {$ref: "#/components/schemas/Organization"}
    PaginatedOrganizations:
      type: object
      properties:
        data: {type: array, items: {$ref: "#/components/schemas/Organization"}}
        meta: {$ref: "#/components/schemas/PageMeta"}
    WorkshopEnvelope:
      type: object
      properties:
        data: {$ref: "#/components/schemas/Workshop"}
    PaginatedWorkshops:
      type: object
      properties:
        data: {type: array, items: {$ref: "#/components/schemas/Workshop"}}
        meta: {$ref: "#/components/schemas/PageMeta"}
    ActivityEnvelope:
      type: object
      properties:
        data: {$ref: "#/components/schemas/Activity"}
    PaginatedActivities:
      type: object
      properties:
        data: {type: array, items: {$ref: "#/components/schemas/Activity"}}
        meta: {$ref: "#/components/schemas/PageMeta"}
...
  responses:
...
    NotFound:
      description: Resource not found in the authorized scope.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    Conflict:
      description: State transition, duplicate, or optimistic concurrency conflict.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    ValidationError:
      description: Input failed schema or business validation.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
    OrganizationItem:
      description: Organization.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/OrganizationEnvelope"}
    OrganizationList:
      description: Organizations.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/PaginatedOrganizations"}
    WorkshopItem:
      description: Workshop.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/WorkshopEnvelope"}
    WorkshopList:
      description: Workshops.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/PaginatedWorkshops"}
    ActivityItem:
      description: Activity.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ActivityEnvelope"}
    ActivityList:
      description: Activities.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/PaginatedActivities"}
...